pyproject.toml
[project]
name = "slack-mcp"
version = "0.1.0"
description = "Slack MCP server for Claude Code slackbot plugin"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "slack-sdk",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slack_mcp"]
